+++ Emuera/Runtime/Script/Statements/Function/Creator.Method.cs
@@ -6954,12 +6954,11 @@ internal static partial class FunctionMethodCreator
 				case "GGETFONTSTYLE":
 					return g.Fontstyle;
 				case "GGETPEN":
-					return g.Pen.Color.ToArgb() & 0xffffffffL;
+					return g.PenColorArgb;
 				case "GGETPENWIDTH":
-					return (long)g.Pen.Width;
+					return g.PenWidth;
 				case "GGETBRUSH":
-					SolidBrush b = (SolidBrush)g.Brush;
-					return b.Color.ToArgb() & 0xffffffffL;
+					return g.BrushColorArgb;
 					#endregion
 			}
 			throw new ExeEE("GraphicsState:" + Name + ":異常な分岐");