+++ Emuera/Runtime/Script/Statements/Instraction.Child.cs
@@ -2414,7 +2414,6 @@ internal sealed partial class FunctionIdentifier
 									if (button.IsButton && button.IsInteger)
 									{
 										count++;
-										goto loopend;
 									}
 								}
 							}
@@ -2517,7 +2516,6 @@ internal sealed partial class FunctionIdentifier
 									if (button.IsButton)
 									{
 										count++;
-										goto loopend;
 									}
 								}
 							}
@@ -2611,7 +2609,6 @@ internal sealed partial class FunctionIdentifier
 									if (button.IsButton && button.IsInteger)
 									{
 										count++;
-										goto loopend;
 									}
 								}
 							}
@@ -2714,7 +2711,6 @@ internal sealed partial class FunctionIdentifier
 									if (button.IsButton)
 									{
 										count++;
-										goto loopend;
 									}
 								}
 							}